+++ src/math.cpp
@@ -15,9 +15,9 @@
 #include <cstdlib>
 
 float getDistance(const sf::Vector2f a, const sf::Vector2f b) {
-  return sqrt(pow(a.x - b.x, 2) + pow(a.y - b.y, 2));
+  return static_cast<float>(sqrt(pow(a.x - b.x, 2) + pow(a.y - b.y, 2)));
 }
 
 float getDistance(const float a, const float b, const float c, const sf::Vector2f point) {
-  return std::abs(a * point.x + b * point.y + c) / sqrt(pow(a, 2)+pow(b, 2));
+  return static_cast<float>(std::abs(a * point.x + b * point.y + c) / sqrt(pow(a, 2)+pow(b, 2)));
 }
